refactor: Migrate daemon_e2e.rs to use shared DaemonTestHarness
Replace the ~400-line DaemonFixture with DaemonTestHarness from
tests/common/mod.rs, eliminating duplicate test infrastructure code.

Changes:
- daemon_e2e.rs now imports DaemonTestHarness via mod common
- Add connect(), rpc_call_with_timeout(), stop_daemon() to harness
- Improve Drop impl to clean up worktrees and use graceful shutdown
- Use module-level #![allow(dead_code)] for shared test utilities
- Keep LiveDaemonFixture (different purpose: live daemon testing)
- Keep daemon_e2e-specific helpers (copy_dir_recursive, etc.)
